advanced collision detection

+++ src/main/java/de/himbiss/ld35/world/Tile.java
@@ -9,16 +9,16 @@ import org.newdawn.slick.opengl.Texture;
  */
 public abstract class Tile implements Renderable {
     public static int TILE_SIZE = 50;
-    public int coordx;
-    public int coordy;
+    public float coordx;
+    public float coordy;
 
     protected String textureKey;
 
-    public int getCoordy() {
+    public float getCoordy() {
         return coordy;
     }
 
-    public int getCoordx() {
+    public float getCoordx() {
         return coordx;
     }
